RPGEditor 프로젝트 로드 실패 수정 - Frames 셀 값에 소수점 좌표 존재
"RPGEditor로 못 불러온다" 재현: ProjectContext.Load()를 직접 호출하는
테스트 하네스로 잡아보니 System.Text.Json이 frames[3][1][1] 같은 셀 값을
Int32로 못 읽고 예외를 던지며 전체 로드가 실패하고 있었다(id=7 애니메이션의
x좌표가 60.5처럼 실제로 소수점을 쓰는 셀이 있음 - 전 데이터셋에 359곳).
MvAnimationData.Frames를 List<List<int[]>>에서 List<List<double[]>>로
변경. 로드/저장/재로드 라운드트립(임시 복사본 대상, 실제 프로젝트 파일은
안 건드림)까지 직접 실행해서 정상 동작 확인함.

+++ RPGEditor/Models/MvAnimationData.cs
@@ -34,8 +34,12 @@ public partial class MvAnimationData : ObservableObject
     [ObservableProperty]
     private int position = 1;
 
-    /// <summary>프레임마다 셀 목록 - 셀 하나는 [pattern, x, y, scale, rotation, mirror, opacity, blendType] 8개 값.</summary>
-    public List<List<int[]>> Frames { get; set; } = [];
+    /// <summary>
+    /// 프레임마다 셀 목록 - 셀 하나는 [pattern, x, y, scale, rotation, mirror, opacity,
+    /// blendType] 8개 값. x/y 등에 소수점 좌표(예: 60.5)가 실제로 존재해서 int[]가
+    /// 아니라 double[]이어야 한다(정수로 읽으면 System.Text.Json이 예외를 던짐).
+    /// </summary>
+    public List<List<double[]>> Frames { get; set; } = [];
 
     public List<MvAnimationTiming> Timings { get; set; } = [];
 }
